test: factor out file server from upgrade test
